+++ src/App.tsx
@@ -11,29 +11,35 @@ import { SettingsScreen } from "./screens/SettingsScreen";
 import { HabitsProvider } from "./state/HabitsContext";
 import { ProjectsProvider } from "./state/ProjectsContext";
 import { GuidesProvider } from "./state/GuidesContext";
+import { CapturesProvider } from "./state/CapturesContext";
+import { ThemeProvider } from "./state/ThemeContext";
 
 function App() {
   return (
-    <HabitsProvider>
-      <ProjectsProvider>
-        <GuidesProvider>
-          <BrowserRouter>
-            <AppShell>
-              <Routes>
-                <Route path="/" element={<HomeScreen />} />
-                <Route path="/habits" element={<HabitsScreen />} />
-                <Route path="/capture" element={<CaptureScreen />} />
-                <Route path="/projects" element={<ProjectsScreen />} />
-                <Route path="/projects/:id" element={<ProjectDetailScreen />} />
-                <Route path="/guides" element={<GuidesScreen />} />
-                <Route path="/guides/:id" element={<GuideDetailScreen />} />
-                <Route path="/settings" element={<SettingsScreen />} />
-              </Routes>
-            </AppShell>
-          </BrowserRouter>
-        </GuidesProvider>
-      </ProjectsProvider>
-    </HabitsProvider>
+    <ThemeProvider>
+      <HabitsProvider>
+        <ProjectsProvider>
+          <CapturesProvider>
+            <GuidesProvider>
+              <BrowserRouter>
+                <AppShell>
+                  <Routes>
+                    <Route path="/" element={<HomeScreen />} />
+                    <Route path="/habits" element={<HabitsScreen />} />
+                    <Route path="/capture" element={<CaptureScreen />} />
+                    <Route path="/projects" element={<ProjectsScreen />} />
+                    <Route path="/projects/:id" element={<ProjectDetailScreen />} />
+                    <Route path="/guides" element={<GuidesScreen />} />
+                    <Route path="/guides/:id" element={<GuideDetailScreen />} />
+                    <Route path="/settings" element={<SettingsScreen />} />
+                  </Routes>
+                </AppShell>
+              </BrowserRouter>
+            </GuidesProvider>
+          </CapturesProvider>
+        </ProjectsProvider>
+      </HabitsProvider>
+    </ThemeProvider>
   );
 }